Commit Inicial

+++ src/controllers/ProdutoController.ts
@@ -1,24 +1,29 @@
 import { Request, Response } from 'express';
-import { ProdutoDAO } from './../dao/ProdutoDAO';
-import { ProdutoModel } from '../domains/ProdutoModel';
+import ProdutoDAO from '../dao/ProdutoDAO';
 
-export class ProdutoController {
+class ProdutoController {
   private _produtoDAO: ProdutoDAO;
 
   constructor() {
     this._produtoDAO = new ProdutoDAO();
   }
 
-  async cadastroItem(req: Request, res: Response) {
-    const { descricao, perecivel } = req.body;
-    const produto = new ProdutoModel({ descricao, perecivel });
-    await this._produtoDAO.cadastroItem(produto);
-    res.status(201).send('Produto Cadastrado com sucesso')
+  async save(req: Request, res: Response) {
+    const id = await this._produtoDAO.save(req.body);
+    res.status(201).json({ id });
   }
 
-  async delete(req: Request, res: Response) {
+  async deleteById(req: Request, res: Response) {
     const { id } = req.params;
-    await this._produtoDAO.delete(id);
-    res.status(204).send('Produto Removido com sucesso')
+    if (!id) return res.status(404).json({ mensagensDeErro: ['ID pendente'] });
+    if (typeof id !== 'number') return res.status(400).json({ mensagensDeErro: ['ID mal formatado'] });
+    const deleted = await this._produtoDAO.delete(id);
+    if (!deleted) {
+      res.status(404).json({ mensagensDeErro: ['Produto não encontrado'] });
+    } else {
+      res.status(201).json({ mensagem: 'Produto deletado com sucesso' });
+    }
   }
 }
+
+export default ProdutoController;
